backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
google-generativeai==0.8.5
gunicorn==23.0.0
python-dotenv==1.1.1
psycopg2==2.9.10
requests==2.32.4
PyJWT==2.10.1
supabase==2.18.0
pillow==12.0.0
gtts==2.5.4
groq==0.32.0
python-multipart==0.0.20
scikit-learn
faiss-cpu
sentence_transformers
pypdf2 
pypdf
tqdm
# === Core scientific libraries ===
numpy==1.26.4
scikit-learn==1.5.0
joblib==1.4.2

# === Sentence embeddings ===
# sentence-transformers==2.7.0

# === Vector search (FAISS) ===
#faiss-cpu==1.8.0

# === Quantum computing ===
qiskit==1.2.2
qiskit-aer==0.14.2

# openai==1.42.0
# python-dotenv
groq